refactor: cleanup relative include paths
