+++ runtime/lib/transport/local_copy/cpu_coprocessor/LocalCpuCoprocessorSession.cpp
@@ -18,7 +18,7 @@
 #include <cstring>
 #include <stdexcept>
 
-#include "../LocalMessageLayout.hpp"
+#include "LocalWireProtocol.hpp"
 
 namespace catalyst::transport::local_copy {
 namespace {
